refactor: rename and prepare to migrate to router
Gonna have to try using Router instead of Swiper for
switching between content. This should be faster on mobile.
And loss of additional fancty functionality is justified.
In future we could add potential nav capabilities in a
more efficient manner.
The approach with router requires for navigation to switch
path for router instead of calling Swiper functions, so
this had to be abstracted for now.